+++ src/Symfony/Component/Validator/ConstraintViolation.php
@@ -32,7 +32,7 @@ class ConstraintViolation implements ConstraintViolationInterface
     /**
      * Creates a new constraint violation.
      *
-     * @param string      $message         The violation message
+     * @param string      $message         The violation message as a string or a stringable object
      * @param string      $messageTemplate The raw violation message
      * @param array       $parameters      The parameters to substitute in the
      *                                     raw violation message
@@ -47,7 +47,7 @@ class ConstraintViolation implements ConstraintViolationInterface
      * @param string|null $code            The error code of the violation
      * @param mixed       $cause           The cause of the violation
      */
-    public function __construct(?string $message, ?string $messageTemplate, array $parameters, $root, ?string $propertyPath, $invalidValue, int $plural = null, $code = null, Constraint $constraint = null, $cause = null)
+    public function __construct($message, ?string $messageTemplate, array $parameters, $root, ?string $propertyPath, $invalidValue, int $plural = null, $code = null, Constraint $constraint = null, $cause = null)
     {
         if (null === $message) {
             @trigger_error(sprintf('Passing a null message when instantiating a "%s" is deprecated since Symfony 4.4.', __CLASS__), E_USER_DEPRECATED);
@@ -56,6 +56,10 @@ class ConstraintViolation implements ConstraintViolationInterface
 
         if (null !== $code && !\is_string($code)) {
             @trigger_error(sprintf('Not using a string as the error code in %s() is deprecated since Symfony 4.4. A type-hint will be added in 5.0.', __METHOD__), E_USER_DEPRECATED);
+        }
+
+        if (!\is_string($message) && !(\is_object($message) && method_exists($message, '__toString'))) {
+            throw new \TypeError('Constraint violation message should be a string or an object which implements the __toString() method.');
         }
 
         $this->message = $message;